Added contribution from github.com/lduchosal to montoouch-bindings
Dear Miguel,

I have build a binding for the DataTrans Credit Card library, that is production ready
and should be available with the other bindings you provide through the
monotouch-binding github.

This binding has been testes and used in real production application
and has been approved in the appstore.

Find here the build project, the sample and the binding.

Thanks for the great monotouch work
Regards
Luc