[chipmunk] unseal Body, Constraint, Shape, ...
Keep a reference of the managed object in the native UserData. Allows
subclasses of Shape and Constraint returned by callbacks to be the right
one instead of parent's.

Allow also user to subclass Bodies, Constraints, Shapes. Arbiters are
still sealed: we don't control their livetimes, so if we keep a
referenced to the managed object, we don't know at which point in time
to clear it.